moved some cli/cmd stuff to lib
